drivers: can: mcp251xfd: Increase max filters and change filter usage type
The mcp251xfd supports upto 32 filters. Also store the filter usage in
uint32_t instead of uint64_t.

+++ drivers/can/can_mcp251xfd.h
@@ -496,7 +496,7 @@ struct mcp251xfd_data {
 	struct mcp251xfd_mailbox mailbox[CONFIG_CAN_MCP251XFD_MAX_TX_QUEUE];
 
 	/* Filter Data */
-	uint64_t filter_usage;
+	uint32_t filter_usage;
 	struct can_filter filter[CONFIG_CAN_MAX_FILTER];
 	can_rx_callback_t rx_cb[CONFIG_CAN_MAX_FILTER];
 	void *cb_arg[CONFIG_CAN_MAX_FILTER];
